Move NavigationStack to its own file for reuse
